Replace lisp_eval_string() with lisp_eval() and changed C interface
Duplicate "catch" code in lisp-eval() is replaced by using evalCatch.

The only interface with the interpreter is now interp->object, which
receives a catch object (error_type message object).

If error_type == nil, object is the result, otherwise object is the err'ed
object (or nil) and message a string object, message->string the error
message.