WIP: implement dynamic memory allocation. lisp_new() doesn't need memory size
The interpreter allocates a minimal chunk of memory to boot itself on
start. After that, every time the Lisp object memory is to small, it is
increased by a specific chunk size.

Both parameters: FLISP_MIN_MEMORY and FLISP_MEMORY_INC_SIZE are defined
lisp.h.

While the approach is working well, there are problems with segfaults, when
using gc_always.  These seem to be related to wrong GC_TRACE'ing

Note: on startup the interpreter prints out, how much memory is used for
the initial startup. We round it up manually to the chunk size.