Re-add facility to specify initial memory
If FLISP_INITIAL_MEMORY is specified, it adds to the minimum amount needed
to start the fLisp interpreter. Please note that only half of the amount is
actually available, the other half is used for garbage collection.

For femto currently 512k(256k) is enough to start up without initiating
garbage collection cycles.