Refactor find-file-noselect, add mode setting and use at startup
Similar to Emacs, find-file-noselect sets up the major mode of the buffer
after creating a file buffer and then runs the find-file-hook.

In our case, we set the syntax highlight mode according to the filename
extension. Instead of the former read-hook method, we use now a property
list stored in the global variable find-file-extension-hightlight-mode.

This way, the user can customize the file extensions/modes in the femto.rc
file.